backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
email-validator==2.1.1
